# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Teacher Auth API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RegisterRequest {
  email: string;
  /** @minLength 6 */
  password: string;
  first_name: string;
  last_name: string;
  phone?: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface UserProfile {
  id: number;
  email: string;
  first_name: string;
  last_name: string;
  phone?: string | null;
  is_active: boolean;
  created_at: string;
}

export interface AuthResponse {
  token: string;
  user: UserProfile;
}

export type CreateTeacherRequestGender =
  (typeof CreateTeacherRequestGender)[keyof typeof CreateTeacherRequestGender];

export const CreateTeacherRequestGender = {
  male: "male",
  female: "female",
  other: "other",
} as const;

export interface CreateTeacherRequest {
  email: string;
  /** @minLength 6 */
  password: string;
  first_name: string;
  last_name: string;
  phone?: string;
  university_name: string;
  gender: CreateTeacherRequestGender;
  year_joined: number;
  subject: string;
  bio?: string;
}

export interface TeacherWithUser {
  id: number;
  user_id: number;
  university_name: string;
  gender: string;
  year_joined: number;
  subject: string;
  bio?: string | null;
  created_at: string;
  user: UserProfile;
}

export interface ErrorResponse {
  error: string;
}