# lib/api-zod/src/generated/types/createTeacherRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Teacher Auth API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateTeacherRequestGender } from "./createTeacherRequestGender";

export interface CreateTeacherRequest {
  email: string;
  /** @minLength 6 */
  password: string;
  first_name: string;
  last_name: string;
  phone?: string;
  university_name: string;
  gender: CreateTeacherRequestGender;
  year_joined: number;
  subject: string;
  bio?: string;
}