# lib/api-zod/src/generated/types/teacherWithUser.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Teacher Auth API
 * OpenAPI spec version: 0.1.0
 */
import type { UserProfile } from "./userProfile";

export interface TeacherWithUser {
  id: number;
  user_id: number;
  university_name: string;
  gender: string;
  year_joined: number;
  subject: string;
  bio?: string | null;
  created_at: Date;
  user: UserProfile;
}